Extract authOptions from route.ts
